pyproject.toml
[project]
name = "cervical-cancer"
version = "0.1.0"
description = "Pap smear cytology inference API"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.125.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "scikit-learn==1.7.2",
    "uvicorn>=0.38.0",

    # CPU-only PyTorch
    "torch==2.2.2+cpu",
    "torchvision==0.17.2+cpu",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
